pyproject.toml
[project]
name = "ai-task-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.3",
    "django-ninja>=1.5.3",
    "psycopg2-binary>=2.9.11",
]
